# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Tomi Wallet
  description: Tomi Wallet Api Documentation
  version: 1.0.0
servers:
  - url: http://localhost:8888
    description: Local server
...
paths:
...
  /collections/groups:
    post: 
      summary: Create Channel
      security:
        - bearerAuth: []
      tags:
        - Groups
      requestBody:
        description: Payload for create chat
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: 
                  type: string
                  example: abc group
                  required: true
                groupImage: 
                  type: string
                  example: http://groupicon.com
                  required: true
                description: 
                  type: string
                  example: this is a group
                groupType: 
                  type: string
                  example: private
                participants: 
                  type: array
                  items:
                    type: object
                    properties:
                      userId:
                        type: string
                        example: 65ae514f564ffea0780ab030               
                subscription:
                  type: object
                  properties:
                    type:
                      type: string
                      example: paid
                    price: 
                      type: string
                      example: 1 
                    interval: 
                      type: string
                      example: 1 month
                    isActive:
                      type: boolean
                      example: true
                dissappearingMessages:
                  type: boolean
                  example: true
                dissappearingMessagesDuration: 
                  type: string
                  example: 24 hours  
      responses:
        201:
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupCreateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        409:
          description: Group already exists
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupAlreadyExist'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
    get:
      summary: Get Groups
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: offset
          schema:
            type: int
          description: page number (start from 1)
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
      tags:
        - Groups
      responses:
        200:
          description: Get Groups Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupListingSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'  
  /collections/groups/pending-requests:      
    get:
      summary: Get Groups
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: offset
          schema:
            type: int
          description: page number (start from 1)
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
      tags:
        - Groups
      responses:
        200:
          description: Get Groups Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupListingSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'    
  /collections/groups/{id}:
    get:  
      summary: fetch Group
      security:
        - bearerAuth: []
      tags:
        - Groups
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: MongoDB ID of the groups to get
      responses:
        201:
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupUpdateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        404:
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupNotFound'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
    patch:  
      summary: Update Group
      security:
        - bearerAuth: []
      tags:
        - Groups
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: MongoDB ID of the groups to update
      requestBody:
        description: Payload for update group
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: 
                  type: string
                  example: abc group
                groupImage: 
                  type: string
                  example: http://groupicon.com
                description: 
                  type: string
                  example: this is a group
                groupType: 
                  type: string
                  example: private          
                subscription:
                  type: object
                  properties:
                    type:
                      type: string
                      example: paid
                    price: 
                      type: string
                      example: 1 
                    interval: 
                      type: string
                      example: 1 month
                    isActive:
                      type: boolean
                      example: true
                dissappearingMessages:
                  type: boolean
                  example: true
                dissappearingMessagesDuration: 
                  type: string
                  example: 24 hours  
      responses:
        201:
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupUpdateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        409:
          description: Group already exists
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupAlreadyExist'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete Group
      security:
        - bearerAuth: []
      tags:
        - Groups 
      parameters:
        - in: path
          name: id
          required: true
          description: Group Id
          schema:
            type: string
            example: 65a9220a28cf8f41efb2a74b
      responses:
        200:
          description: Group Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ChatDeleteSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
  /collections/groups/{id}/leave:
    post: 
      summary: Leave Group
      security:
        - bearerAuth: []
      tags:
        - Groups
      parameters:
        - in: path
          name: id
          required: true
          description: Group Id
          schema:
            type: string
            example: 65a9220a28cf8f41efb2a74b
      responses:
        200:
          description: Group left successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupUpdateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        404:
          description: Group not Found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupNotFound'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
  /collections/groups/{id}/join:
    post: 
      summary: Join Group
      security:
        - bearerAuth: []
      tags:
        - Groups
      parameters:
        - in: path
          name: id
          required: true
          description: Group Id
          schema:
            type: string
            example: 65a9220a28cf8f41efb2a74b
      requestBody:
        description: Payload for adding user to groups
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subscription:
                  type: object
                  properties:
                    status:
                      type: string
                      example: active
                    startedAt: 
                      type: string
                      example: "2023-10-01T00:00:00Z"
                    expiresAt: 
                      type: string
                      example: "2024-10-01T00:00:00Z"
                    paymentHash: 
                      type: string
                      example: "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
      responses:
        200:
          description: Group Joined successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupUpdateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        404:
          description: Group not Found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupNotFound'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
  /collections/groups/{id}/transfer-ownership:
    patch:  
      summary: Update Group Owner
      security:
        - bearerAuth: []
      tags:
        - Groups
      requestBody:
        description: Payload for update group owner
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: 
                  type: string
                  example: asd3c234sdasd234asd
      responses:
        201:
          description: Group Admin updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupUpdateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        404:
          description: Group not Found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupNotFound'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
  /collections/groups/{id}/mute:
    patch:
      summary: Mute Group
      security:
        - bearerAuth: []
      tags:
        - Groups 
      parameters:
        - in: path
          name: id
          required: true
          description: Group Id
          schema:
            type: string
            example: 65a9220a28cf8f41efb2a74b
      requestBody:
        description: Payload for react to message
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isMuted:
                  type: string
                  example: false
                duration:
                  type: string
                  example: "8h | 1w | always"
      responses:
        200:
          description: Group Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupCreateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
  /collections/groups/{id}/status:
    patch:
      summary: Update Group Message status
      security:
        - bearerAuth: []
      tags:
        - Groups 
      parameters:
        - in: path
          name: id
          required: true
          description: Group Id
          schema:
            type: string
            example: 65a9220a28cf8f41efb2a74b
      requestBody:
        description: Payload for react to message
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: read
      responses:
        200:
          description: Group Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupCreateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
  /collections/groups/{id}/clear:
    patch:
      summary: Clear Group
      security:
        - bearerAuth: []
      tags:
        - Groups 
      parameters:
        - in: path
          name: id
          required: true
          description: Chat Id
          schema:
            type: string
            example: 65a9220a28cf8f41efb2a74b
      responses:
        200:
          description: Group Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupCreateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
  /collections/groups/{id}/accept-ownership:
    patch:  
      summary: Update Group Owner
      security:
        - bearerAuth: []
      tags:
        - Groups
      requestBody:
        description: Payload for update group owner
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: 
                  type: string
                  example: asd3c234sdasd234asd
      responses:
        201:
          description: Group Admin updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupUpdateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        404:
          description: Group not Found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupNotFound'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
  /collections/groups/{id}/participants:
    get:
      summary: Get Group Participants
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: offset
          schema:
            type: int
          description: page number (start from 1)
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
      tags:
        - Groups
      responses:
        200:
          description: Get Group Participants Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/GroupParticipantsListingSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'   
...
  /collections/groups/messages:
    post: 
      summary: Send Message
      security:
        - bearerAuth: []
      tags:
        - Group Messages
      requestBody:
        description: Payload for send message
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                groupId: 
                  type: string
                  example: 65a9220a28cf8f41efb2a74b
                replyTo: 
                  type: string
                  example: 65a9220a28cf8f41efb2a74b
                content: 
                  type: object
                  required: true
                  properties:
                    body:
                      type: string
                      example: Hello, how are you?
                    type: 
                      type: string
                      example: text
                    metaData: 
                      type: object
                viewType:
                  type: string
                  example: standard || one-time-view
                messageAccess:
                  type: object
                  properties:
                    type: 
                      type: string
                      example: free
                    isLocked: 
                      type: boolean
                      example: false
                    price: 
                      type: number
                      example: 0.01
                    currency: 
                      type: string
                      example: ETH
                    metaData: 
                      type: object                
      responses:
        201:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/MessageCreateSuccess'
        400:
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError'
    get:
      summary: Get Group Messages
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: nextCursor
          schema:
            type: string
          description: the id of last message 
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
        - in: query
          name: groupId
          schema:
            type: string
          description: chatId to get messages
        - in: query
          name: types
          schema:
            type: string
          description: string of array of types to filter data
      tags:
        - Group Messages
      responses:
        200:
          description: Get Group Messages Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ChatMessagesListingSuccess'
        401:
          description: UnAuthorized (Access Token expired or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UnAuthorized'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerError' 
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      summary: Bad Request (Api Payload validation failed)
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Validation Failed
        details:
          type: array
          items:
            type: object
            properties:
              "{payload_field}":
                type: string
                example: "{payload_field validation error message}"
    ServerError:
      summary: Internal Server Error
      type: object
      properties:
        statusCode:
          type: integer
          example: 500
        message:
          type: string
          example: Something went wrong
        details:
          example: null
    UnAuthorized:
      summary: UnAuthorized
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Jwt expired/invalid
        details:
          example: null
...
    GroupCreateSuccess:
      summary: Group Create Success
      type: object
      properties:
        statusCode: 
          type: number
          example: 201
        message: 
          type: string
          example: Group created successfully
        data: 
          type: object
          properties: 
            name: 
              type: string
              example: abc
            groupImage: 
              type: string
              example: gsdg
            description: 
              type: string
              example: sdf
            owner: 
              type: string
              example: 67ee868047458f63106862b5
            groupType: 
              type: string
              example: public
            groupMembersCount: 
              type: number
              example: 1
            subscription: 
              type: object
              properties: 
                type: 
                  type: string
                  example: free
                price: 
                  type: number
                  example: 0
                interval: 
                  type: string
                  format: nullable
                isActive: 
                  type: boolean
            dissappearingMessages: 
              type: boolean
            dissappearingMessagesDuration: 
              type: string
              format: nullable
            _id: 
              type: string
              example: 680f288feb9fea929827c0b4
            lastUpdated: 
              type: string
              format: date-time
              example: 2025-04-28T07:04:47.280Z
            createdAt: 
              type: string
              format: date-time
              example: 2025-04-28T07:04:47.281Z
            updatedAt: 
              type: string
              format: date-time
              example: 2025-04-28T07:04:47.281Z
            __v: 
              type: number
              example: 0
    GroupAlreadyExist:
      summary: Group Already Exist
      type: object
      properties:
        statusCode:
          type: number
          example: 409
        message:
          type: string
          example: Group already exists
        details:
          type: string
          format: null   
    GroupListingSuccess:
      summary: Group Listing Success
      type: object
      properties:
        statusCode: 
          type: number
          example: 200
        message: 
          type: string
          example: Group Requests
        data: 
          type: object
          properties: 
            groups: 
              type: array
              items: 
                type: object
                properties: 
                  _id: 
                    type: string
                    example: 680f288feb9fea929827c0b4
                  name: 
                    type: string
                    example: abc
                  groupImage: 
                    type: string
                    example: gsdg
                  description: 
                    type: string
                    example: sdf
                  owner: 
                    type: string
                    example: 67ee868047458f63106862b5
                  groupType: 
                    type: string
                    example: public
                  groupMembersCount: 
                    type: number
                    example: 1
                  subscription: 
                    type: object
                    properties: 
                      type: 
                        type: string
                        example: free
                      price: 
                        type: number
                        example: 0
                      interval: 
                        type: string
                        format: nullable
                      isActive: 
                        type: boolean
                  dissappearingMessages: 
                    type: boolean
                  dissappearingMessagesDuration: 
                    type: string
                    format: nullable
                  lastUpdated: 
                    type: string
                    format: date-time
                    example: 2025-04-28T07:04:47.280Z
                  createdAt: 
                    type: string
                    format: date-time
                    example: 2025-04-28T07:04:47.281Z
                  updatedAt: 
                    type: string
                    format: date-time
                    example: 2025-04-28T07:04:47.281Z
                  __v: 
                    type: number
                    example: 0
            pages: 
              type: number
              example: 1
            count: 
              type: number
              example: 1
...
    GroupNotFound:
      summary: Group Does Not Exists
      type: object
      properties:
        statusCode:
          type: number
          example: 404
        message:
          type: string
          example: Group not Found
        details:
          type: string
          format: null
...